src/utils/data_converter_2017/requirements.txt
mysql-connector-python-rf==2.2.2
pydgraph==20.3.0
